Add files via upload
Finished insights which now has two more tables (We only need one more now)

Enterprise is set to have a form so that's one form down out of the 3 ones we need.

We already have external links so that's checked off, and I also added an email link too in insights so that is ALSO checked off. This is getting close to a finish line.